+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Changes
 
+## Unreleased
+- [Allow proj-sys to link to `proj_d.lib`. This prevents a linker error when building in debug mode with MSVC](https://github.com/georust/proj/pull/83)
+
 ## 0.22.0
 - Update PROJ to 7.2.1 via proj-sys 0.19.0